==> 01_Daily_diary_data.Rmd <==
---
title: "Daily diary data"
output: html_notebook
---

```{r Github}
# usethis::create_from_github("https://github.com/Sfglatt/Long_Covid_trials.git",
#                            destdir = "Github/Long_Covid_trials")
```

```{r Packages}
# Update the packages needed
if (!require("cowplot")) {install.packages("cowplot"); require("cowplot")}               
if (!require("DescTools")) {install.packages("DescTools"); require("DescTools")}          
if (!require("dplyr")) {install.packages("dplyr"); require("dplyr")}                      
if (!require("dplyr")) {install.packages("dplyr"); require("dplyr")}                      
if (!require("htmlwidgets")) {install.packages("htmlwidgets"); require("htmlwidgets")}    
if (!require("kableExtra")) {install.packages("kableExtra"); require("kableExtra")}       
if (!require("lme4")) {install.packages("lme4"); require("lme4")}                        
if (!require("moments")) {install.packages("moments"); require("moments")}               
if (!require("naniar")) {install.packages("naniar"); require("naniar")}                   
if (!require("plotly")) {install.packages("plotly"); require("plotly")}                  
if (!require("psych")) {install.packages("psych"); require("psych")}                      
if (!require("pwr")) {install.packages("pwr"); require("pwr")}                           
if (!require("readxl")) {install.packages("readxl"); require("readxl")}                   
if (!require("rmcorr")) {install.packages("rmcorr"); require("rmcorr")}                   
if (!require("sjPlot")) {install.packages("sjPlot"); require("sjPlot")}                   
if (!require("tidyr")) {install.packages("tidyr"); require("tidyr")}                      
if (!require("tidyverse")) {install.packages("tidyverse"); require("tidyverse")}          
```

```{r Data}
# Import data 
EMA_dat <- read.csv("Raw_data/Long_COVID_EMA_Data_Final_Burst_De-Identified.csv") 

# Look at the data
head(EMA_dat)
glimpse(EMA_dat)
EMA_dat_1 <- EMA_dat[-c(1:2), ] # remove first two rows 

# Rename the ID and start date columns to be consistent with the other datasets
EMA_dat_1 <- EMA_dat_1 %>%
  dplyr::rename(PID = RecipientLastName,
         StartDate = RecordedDate)

# Rearrange the dataset by participant and time 
EMA_dat_1_sorted <- EMA_dat_1 %>%
  dplyr::arrange(PID, StartDate)

View(EMA_dat_1_sorted)

# the first three rows don't not have an ID (*preview*), so exclude those
EMA_dat_1_sorted <- EMA_dat_1_sorted[-c(1:3), ]
```

```{r Adherence and missingness}
# Change PIDs to just numeric 
EMA_dat_1_sorted <- EMA_dat_1_sorted %>% mutate(PID = as.numeric(sub("^P", "", PID)))

# Convery characters to numeric
EMA_dat_1_sorted[, 8:37] <- lapply(EMA_dat_1_sorted[, 8:37], as.numeric)

# Look at participant adherence
(days_per_participant <- EMA_dat_1_sorted %>%
  dplyr::group_by(PID) %>%
  dplyr::summarise(days_count = n_distinct(StartDate))) # 4 participants have 20/21 days, 1 has 5.

# Look at missingness
# Change blank cells to NA
EMA_dat_1_sorted_processed <- EMA_dat_1_sorted %>%
  mutate(across(.cols = 8:38, ~ifelse(. == "", NA, .)))

# Summarize missing values (proportions) for each variable per participant from filled out diaries
(missing_summary_1 <- EMA_dat_1_sorted_processed %>%
  dplyr::group_by(PID) %>%
  dplyr::summarise(across(.cols = 8:38, ~mean(is.na(.)), .names = "prop_{.col}")))

# Summarize missing values for each variable per participant out of 21 days 
(missing_summary_2 <- EMA_dat_1_sorted_processed %>%
  dplyr::group_by(PID) %>%
  dplyr::summarise(across(.cols = 8:38, ~sum(is.na(.)) / 21, .names = "prop_{.col}")))

write.csv(EMA_dat_1_sorted_processed, "01_Output/Burst_three_EMA_data_sorted.csv")
```

```{r Baseline}
# Add trial baseline and demographic data
EMA_baseline_dat <- read.csv("Raw_data/Cohort_3_Baseline_Data_Nonnumeric.csv")
head(EMA_baseline_dat)

EMA_baseline_dat <- EMA_baseline_dat[-1, ]

# Rename and reformat ID column to match daily diary data
EMA_baseline_dat <- EMA_baseline_dat %>%
  dplyr::rename(PID = ID)

EMA_baseline_dat$PID <- gsub("^P", "", EMA_baseline_dat$PID)

# Rename all columns to show they're from baseline assessment
EMA_baseline_dat <- EMA_baseline_dat %>%
  rename_with(~ paste0("Baseli_", .), .cols = -PID)

# Merge
merged_data <- EMA_dat_1_sorted_processed %>%
  left_join(EMA_baseline_dat, by = "PID")

# Remove baseline data from *every* daily diary row per participant (i.e., only the first row per participant should also have basline)
merged_data <- merged_data %>%
  dplyr::group_by(PID) %>%
  dplyr::mutate(across(starts_with("Baseli_"), 
                       ~ if_else(row_number() == 1, ., NA_character_), 
                       .names = "{.col}")) %>%
  ungroup()

View(merged_data)

write.csv(merged_data, "01_Output/Burst_three_EMA_data_sorted_with_baselines.csv")

# brief demos
merged_data$Baseli_age <- as.numeric(merged_data$Baseli_age)
summary(merged_data$Baseli_age, na.rm = TRUE)
table(merged_data$Baseli_birth_sex)
table(merged_data$Baseli_ethnicity)
```

```{r Variability}
# Looking at between / within person variance for some example items with intraclass correlation coefficients
uncon_coping_1 <- lmer(coping_1 ~ 1 + (1 | PID), data = EMA_dat_1_sorted)
tab_model(uncon_coping_1, p.val = "kr") 

uncon_coping_2 <- lmer(coping_2 ~ 1 + (1 | PID), data = EMA_dat_1_sorted)
tab_model(uncon_coping_2, p.val = "kr") 

uncon_coping_3 <- lmer(coping_3 ~ 1 + (1 | PID), data = EMA_dat_1_sorted)
tab_model(uncon_coping_3, p.val = "kr") 

uncon_coping_4 <- lmer(coping_4 ~ 1 + (1 | PID), data = EMA_dat_1_sorted)
tab_model(uncon_coping_4,  p.val = "kr") 

uncon_coping_5 <- lmer(coping_5 ~ 1 + (1 | PID), data = EMA_dat_1_sorted)
tab_model(uncon_coping_5)

uncon_qol <- lmer(qol ~ 1 + (1 | PID), data = EMA_dat_1_sorted)
tab_model(uncon_qol) \
```

```{r Long form data}
# sorting the data into long format: 

# Change column forms
EMA_dat_1_sorted <- EMA_dat_1_sorted %>% mutate(across(-c(PID, StartDate), as.character))

# Reshape dataframe to long format
long_df <- EMA_dat_1_sorted %>% pivot_longer(cols = -c(PID, StartDate), 
               names_to = "item", 
               values_to = "answer_id")

# Change PIDs to just numeric 
long_df <- long_df %>% mutate(PID = as.numeric(sub("^P", "", PID)))

# Change answer_id to numeric 
long_df <- long_df %>% mutate(answer_id = as.numeric(answer_id))

# save the long data
write.csv(long_df, "01_Output/Burst_three_EMA_data_long_form.csv")

# variables to include 
ema_mh_vars <- c("pre_energy", "energy", "post_energy", "qol", "hopelessness", "health", "restedness", "health",
                 "falling_asleep", "staying_asleep", "waking_early", "isolation", "busy", "anxiety", 
                 "depression", "lc_sx_1", "lc_sx_2", "lc_sx_3", "lc_sx_4", "lc_sx_5", "lc_sx_6", "lc_sx_7",
                 "lc_sx_8", "coping_1", "coping_2", "coping_3", "coping_4", "coping_5","functioning")

# names(long_df)  
# str(long_df)  

# calculate individual mean, sd, skew, etc. 
(df_desc <- long_df %>%
  dplyr::filter(item %in% ema_mh_vars) %>%  
  dplyr::group_by(PID, item) %>%
  dplyr::summarize(
    imean = mean(answer_id, na.rm = TRUE),
    imed = median(answer_id, na.rm = TRUE),
    isd = sd(answer_id, na.rm = TRUE),
    iskew = moments::skewness(answer_id, na.rm = TRUE),
    irmssd = psych::rmssd(answer_id, na.rm = TRUE)
  ))

# view individual mean, sd, skew, etc. 
t1 <- df_desc %>%
  dplyr::filter(item %in% ema_mh_vars) %>%
  ungroup() %>%
  dplyr::group_by(item) %>%
  dplyr::select(-PID) %>%
  dplyr::filter(!is.nan(iskew)) %>%
  dplyr::summarize(across(
    everything(),
    list(mean = mean, sd = sd),
    .names = "{.col}_{.fn}"
  )) %>%
  dplyr::group_by(item) %>%
  dplyr::summarize(across(
    everything(),
    ~ round(., digits = 2)
  )) %>%
  mutate(
    iMean = paste0(format(imean_mean, drop0trailing = FALSE), " (", format(imean_sd, drop0trailing = FALSE), ")"),
    iMedian = paste0(format(imed_mean, drop0trailing = FALSE), " (", format(imed_sd, drop0trailing = FALSE), ")"),
    iSD = paste0(format(isd_mean, drop0trailing = FALSE), " (", format(isd_sd, drop0trailing = FALSE), ")"),
    iSkew = paste0(format(iskew_mean, drop0trailing = FALSE), " (", format(iskew_sd, drop0trailing = FALSE), ")"),
    iRMSSD = paste0(format(irmssd_mean, drop0trailing = FALSE), " (", format(irmssd_sd, drop0trailing = FALSE), ")")
  ) %>%
  dplyr::select(item, iMean, iMedian, iSD, iSkew, iRMSSD) %>%
  ungroup()

knitr::kable(t1, format = "markdown", align = "c")

# Save individuals' descriptives
write.csv(df_desc, "01_Output/Burst_three_Individual_item_descriptives.csv" )
```

```{r Participant average and SD visualization}
# Plot the mean for each individuals' variable endorsement (e.g., participant 15's average hopefulness score).
# Pre-definition so that this will plot consistently across different participants 
table(EMA_dat_1$PID)

numeric_cols <- c("pre_energy", "energy", "post_energy", "qol", "hopefulness", "health", "restedness", 
                  "falling_asleep", "staying_asleep", "waking_early", "isolation", "busy", "anxiety", 
                  "depression", "lc_sx_1", "lc_sx_2", "lc_sx_3", "lc_sx_4", "lc_sx_5", "lc_sx_6", 
                  "lc_sx_7", "lc_sx_8", "coping_1", "coping_2", "coping_3", "coping_4", "coping_5",
                  "functioning") # this will be used to define what to plot, and separately, for factors/levels. 

palette_28 <- c("#00CC33", "#006600", "#009E73", "#00FFCC", "#00FFFF", "#33CCCC", "#56B4E9",
              
                "#006666", "#3333FF", "#003399", "#FF33CC", "#FF6699", "#FF0066", "#FF6666",
              
                "#996699", "#993366", "#990099", "#9933CC", "#663399", "#330066", "#CC3333",
              
                "#E69F00", "#FFCC00", "#FFCC99", "#FFCCFF", "#FF6600", "#FF9933", "#CC9966")


#### Individual participant Variable means - GROUP plot ###

plot_all_ind_means <- long_df %>%
  group_by(PID, item) %>%
  summarise(imean = mean(answer_id, na.rm = TRUE),
            imedian = median(answer_id, na.rm = TRUE),
            isd = sd(answer_id, na.rm = TRUE)) %>%
  ungroup() %>%
  filter(item %in% numeric_cols) %>%
  mutate(item = as.factor(item)) %>%
  # mutate(item = fct_reorder(item, imean, .fun = "median", .na_rm = TRUE)) %>%
   # It often is meaningful to plot in order of the median, which the above comment does. 
   # I did not do that and opted to choose  one order so that the same order can be followed for other plots. 
  mutate(item = factor(item, levels = numeric_cols)) %>%
  ggplot(aes(x = as.numeric(item), y = imean, shape = as.factor(PID))) +
  geom_point(aes(fill = item, color = item, 
                 text = paste("iSD: ", round(isd, 2), "<br>iMedian: ", round(imedian, 2))), 
             position = position_jitter(width = 0.2), size = 2) +
  scale_fill_manual(values = palette_28) +
  scale_color_manual(values = palette_28) +
  scale_shape_manual(values = c(4, 15, 16, 17, 18)) +  
  scale_y_continuous(breaks = seq(0, 10, 1)) +  
  scale_x_continuous(breaks = seq_along(unique(long_df$item)), labels = seq_along(unique(long_df$item))) +  
  guides(
    shape = guide_legend(title = "Participants", 
                         override.aes = list(shape = c(4, 15, 16, 17, 18), 
                                             labels = c("P15", "P16", "P19", "P20", "P21"))))
                                             # labels = c("P4", "P15", "P16", "P17", "P18")
                                             

(plot_all_ind_means_inter <- ggplotly(plot_all_ind_means))

htmlwidgets::saveWidget(ggplotly(plot_all_ind_means_inter),
                        file = file.path("Long_Covid_trials",                                                     
                                         "Interactive_means_all_vars.html"))
```

```{r Daily series per participant and variable}
# This chunk is plotting each participant (IDs 15, 16, 19, and 21) daily diary and arranging them 2x2 for each variable. (e.g., a PDF with daily "hopelessness" for each participant.) 

output_dir <- "01_Output"

numeric_cols_1 <- c("qol", "hopefulness", "health", "restedness", 
                    "falling_asleep", "staying_asleep", "waking_early", 
                    "isolation", "busy",
                    "anxiety", "depression", 
                    "lc_sx_1", "lc_sx_2", "lc_sx_3", "lc_sx_4", "lc_sx_5", "lc_sx_6", "lc_sx_7", "lc_sx_8",
                    "coping_1", "coping_2", "coping_3", "coping_4", "coping_5", "functioning")

numeric_cols_2 <- c("pre_energy", "energy", "post_energy")

participants_of_interest <- c(15, 16, 19, 21)

palette <- c("#FF6699", "#FF9966", "#CC99CC")

variable_colors <- setNames(palette, unique(numeric_cols_2))

for (variable_of_interest in numeric_cols_2) {
  variable_plots <- list()
  
  for (pid in participants_of_interest) {
    participant_data <- long_df %>%
      filter(PID == pid, item == variable_of_interest) %>%
      mutate(DayNumber = as.integer(factor(StartDate, levels = unique(StartDate))))
    
    min_value <- floor(min(na.omit(participant_data$answer_id)))
    max_value <- ceiling(max(na.omit(participant_data$answer_id)))
    
    if (!is.finite(min_value)) min_value <- 0
    
    breaks <- seq(0, 10, by = 1)
    limits <- c(0, 10)
    
    plot <- ggplot(participant_data, aes(x = DayNumber, y = answer_id, group = item, color = item)) +
      geom_line(alpha = 0.7, size = 1.5) +
      geom_vline(xintercept = c(7, 14, 21), linetype = "dashed", color = "grey") +  
      scale_color_manual(values = variable_colors) +  
      theme_minimal() +
      labs(
        x = "Days",
        y = "Endorsement",
        title = paste("Series for Participant", pid, "-", variable_of_interest)
      ) +
      scale_y_continuous(breaks = breaks, limits = limits)
    
    variable_plots[[as.character(pid)]] <- plot
  }
  
  arranged_plot <- cowplot::plot_grid(plotlist = variable_plots, ncol = 2)
  
  arranged_plot_filename <- file.path(output_dir, paste("Series_", 
                                                        variable_of_interest, 
                                                        "_arranged.pdf",
                                                        sep = ""))
  
  ggsave(arranged_plot_filename, arranged_plot, width = 12, height = 10, units = "in")
}
```

```{r Repeated measures correlations}
# minimum detectable effect size? Function for repeated measure correlations:
power.rmcorr <- function(k, N, effectsizer, sig) {
    pwr.r.test(n = ((N)*(k-1))+1, r = effectsizer, sig.level = sig) } 

power.rmcorr(19,   # effective n of 77
             4,    # excluding P20
             .33,  # random
             .05)  # sig.

# correlations
rmcorr <- rmcorr_mat(PID, 
                     variables = c("qol", "hopefulness", "health", "restedness",
                                   "falling_asleep",  "staying_asleep", "waking_early",
                                   "isolation", "busy", "anxiety", "depression",
                                   "lc_sx_1", "lc_sx_2", "lc_sx_3", #"lc_sx_4", 
                                   "lc_sx_5", "lc_sx_6", "lc_sx_7", "lc_sx_8", 
                                   "coping_1", "coping_2", "coping_3", "coping_4", "coping_5", 
                                   "functioning", "energy"), 
                     subset(EMA_dat_1_sorted, PID != 20), 
                     CI.level = 0.95)

rmcorr_summary <- rmcorr$summary

write.csv(rmcorr_summary, "01_Output/Daily_correlations_2.csv")

p.vals <- rmcorr$summary$p.vals
(p.vals.bonferroni <- p.adjust(p.vals, method = "bonferroni", n = length(p.vals)))
```
